Finalized Settings Framework (also other stuff!)
This commit does a TON of stuff.  Settings page works from the new sexy
ActionBar.  Settings are persistent, and both Availability and Radius
are set-able (Still not quite sure where it's saved, but it's functional
for now!).  Made the app look prettier for demonstration by aligning all
buttons to the center of the screen.  Etc.